ElectronLiquidRG

Documentation for ElectronLiquidRG.

ElectronLiquidRG.KO

ElectronLiquidRG.c_coeff_phe

ElectronLiquidRG.c_coeff_pp

ElectronLiquidRG.vertex3

ElectronLiquidRG.∂R_∂F_exchange

ElectronLiquidRG.∂R_∂Λ_exchange

ElectronLiquidRG.KO — Function

function KO(para::ParaMC, kamp=para.kF, kamp2=para.kF; N=100, mix=1.0, verbose=1, ct=false)

Calculate Fs with the self-consistent equation
```math
f_s = -\left< (v_q+f_s)/(1-(v_q+f_s)Π_0) \right>
```

ElectronLiquidRG.c_coeff_phe — Function

uPi_0u particle-hole exchange diagram.

(up, up, up, up): the internal bubble is spin down, and there is a Fermi loop contributing a minus sign

(up, up, down, down): the internal bubble is up and down, and there is no closed Fermi loop, so there is no additional sign.

For a reference of the diagram sign, according to ElectronLiquid.Ver4.MC_PH subroutine, for rs=1, mass=10.0, Yukawa interaction

order upup updown

1 (exchange)        0.152(1)        0.0 
2 (PHEr)            0.0088(1)    -0.0119(2)
2 (PPr)             0.002083     0.02198(42)

WARNING 1: Following the convention of the RG application, we need to add an additional minus to the above results. WARNING 2: The main contributions in the PPr diagrams are from the UV, so it should be compared with Λ∞ - uG0G0u where Λ∞ needs to be fitted.

ElectronLiquidRG.c_coeff_pp — Function

-uG0G0*u + #Λ∞ ladder diagram subtracting the constant term. The (up, up, up, up) is zero and (up, up, down, down) is non-zero

ElectronLiquidRG.vertex3 — Method

function vertex3(para::ParaMC;

calculate particle-particle-pair * R. Only spin (up, up; down, down) is non-zero!

In the large kamp limit, with kamp2 =0, it approaches to -me^2NF*pi/2/kamp

ElectronLiquidRG.∂R_∂F_exchange — Method

function ∂Rs_∂Fs_exchange(paras, Λgrid; ct=false)

 return N_F ∂<R_{k_1-k_2}>/∂Fs

ElectronLiquidRG.∂R_∂Λ_exchange — Method

function ∂Rs_∂Λ_exchange(paras, Λgrid; ct=false)

 return N_F ∂<R_{k_1-k_2}>/∂Λ